customer_order_service/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8,<3.0
django-allauth>=0.48.0
python-decouple>=3.6
python-dotenv>=0.19.2
requests>=2.25.0
djangorestframework
africastalking